+++ index.html
@@ -37,5 +37,7 @@
       </div>
   </div>
   </footer>
-  <script type="text/javascript" src="spotify.js"></script>
+  <script type="text/javascript" src="info.js"></script>
+  <script type="text/javascript" src="control.js"></script>
+  <script type="text/javascript" src="interface.js"></script>
 </html>
