+++ info.js
@@ -1,65 +1,10 @@
 var $ = require('jquery');
-const remote = require('electron').remote;
-var fadeTime = 500;
-var updateMs = 1000;
-var changeMs = 200;
-var isControl = false;
-var timerId;
-
-function control(type) {
-    isControl = true;
-    $.ajax({
-        async: true,
-        url: 'http://localhost:8080/control',
-        data: {
-            type: type
-        },
-        type: 'GET',
-        success: function() {
-            update(true);
-            setTimeout(function() {
-                update(true);
-            }, 100);
-            setTimeout(function() {
-                update(true);
-            }, 200);
-            setTimeout(function() {
-                isControl = false;
-            }, 300);
-        }
-    });
-};
-
-
-
-function volume_control(value) {
-    var  throttleFunction  =  function (func, delay) {
-        if (timerId) {
-            return;
-        }
-        timerId  =  setTimeout(function () {
-            func();
-            timerId  =  undefined;
-        }, delay)
-    };
-    throttleFunction(function() {
-        $.ajax({
-            url: 'http://localhost:8080/volume',
-            type: 'GET',
-            data: {
-                value: value
-            },
-            success: function() {
-                console.log('volume set to ', value)
-            }
-        });
-    }, 200);
-};
 var updateInt;
+var updateMs = 1000;
 var mySong;
 var myArtist;
 var myAlbum;
-var isPlaying;
+
 // Fades first text, only called once
 function start(){
     $.ajax({
@@ -168,12 +113,8 @@ function update(CONTROL){
                                 else {
                                     document.getElementById("bg").innerHTML = `<img src="${data.body.item.album.images[0].url}">`;
                                 }
-                                fadeIn();  
-                                function timerMethod() {
-                                    if(runCount > 1) clearInterval(timerId);
-                                    update(false);
-                                };
-                                var timerId = setInterval(timerMethod, 100);
+                                fadeIn();
+                                setTimeout(update, 200);
                             }, fadeTime - 200) // For lag
                         };
                         mySong = data.body.item.name;
@@ -221,198 +162,3 @@ function update(CONTROL){
         }
     });
 };
-
-// Toggle playback
-var singleClick = function(){
-    $.ajax({
-        async: true,
-        url: 'http://localhost:8080/currently-playing',
-        type: 'GET',
-        success: function(data) {
-            isPlaying = data.body.is_playing;
-            // Play if paused; pause if playing
-            if (isPlaying == true){
-                control('pause');
-                $('#toggle').removeClass().addClass('fa fa-play');
-                console.log('Pausing music');
-            }
-            else if (isPlaying == false){
-                control('play');
-                $('#toggle').removeClass().addClass('fa fa-pause');
-                console.log('Playing music');
-            }
-            else {
-                console.log('No track loaded');
-            };
-        }
-    });
-};
-
-// Get mouse x position
-function xCoords(event) {
-    var x = event.clientX;
-    return x;
-};
-// Get mouse y position
-function yCoords(event) {
-    var y = event.clientY;
-    return y;
-};
-// Skip to next song in queue
-var doubleClickFwd = function(){
-    console.log('Skipping forward');
-    control('forward');
-    fadeOut();
-    setTimeout(function() {
-        update(false);
-    }, changeMs);
-};
-// Skip back to previous song
-var doubleClickBkwd = function(){
-    console.log('Skipping backward');
-    control('backward');
-    fadeOut();
-    setTimeout(function() {
-        update(false);
-    }, changeMs);
-};
-
-var isFullscreen = false;
-
-document.addEventListener('webkitfullscreenchange', function () {
-    console.log('full change')
-    isFullscreen = !!document.webkitIsFullScreen;
-}, false);
-
-function fullScreen(elem) {
-    // ## The below if statement seems to work better ## if ((document.fullScreenElement && document.fullScreenElement !== null) || (document.msfullscreenElement && document.msfullscreenElement !== null) || (!document.mozFullScreen && !document.webkitIsFullScreen)) {
-    if (!isFullscreen) {
-        if (elem.webkitRequestFullScreen) {
-            elem.webkitRequestFullScreen(Element.ALLOW_KEYBOARD_INPUT);
-        }
-    } else {
-        if (document.webkitCancelFullScreen) {
-            document.webkitCancelFullScreen();
-        }
-    };
-};
-var firing = false;
-$("#toggle").click(function() {
-    if(firing)
-        return;
-        singleClick();
-    firing = false;
-});
-
-$("#seek").click(function () {
-    if(firing)
-        return;
-    doubleClickFwd();
-    firing = false;
-});
-
-$("#previous").click(function () {
-    if(firing)
-        return;
-    doubleClickBkwd();
-    firing = false;
-});
-
-$("#full").click(function () {
-    if(firing)
-        return;
-    var elem = document.body;
-    fullScreen(elem);
-})
-
-$("#x").click(function () {
-    if(firing)
-        return;
-    var window = remote.getCurrentWindow();
-    window.close();
-})
-
-$("#square").click(function () {
-    if(firing)
-        return;
-    var window = remote.getCurrentWindow();
-    if (!window.isMaximized()) {
-        window.maximize();          
-    } else {
-        window.unmaximize();
-    }
-})
-
-$("#minimize").click(function () {
-    if(firing)
-        return;
-    var window = remote.getCurrentWindow();
-    window.minimize(); 
-})
-
-// Hide mouse function
-$(document).ready(function() {
-    var idleMouseTimer;
-    var forceMouseHide = false;
-    $("body").css('cursor', 'none');
-    hideHeader();
-    $("body").mousemove(function() {
-        if(!forceMouseHide) {
-            $("body").css('cursor', '');
-            showHeader();
-            clearTimeout(idleMouseTimer);
-            idleMouseTimer = setTimeout(function() {
-                $("body").css('cursor', 'none');
-                hideHeader();
-                forceMouseHide = true;
-                setTimeout(function() {
-                    forceMouseHide = false;
-                }, 200);
-            }, 2500);
-        };
-    });
-});
-
-// Fading functions
-function fadeIn() {
-    $('#bg').fadeIn(fadeTime);
-    setTimeout(function(){
-        $('h1').fadeIn(fadeTime);
-        $('h2').fadeIn(fadeTime);
-    }, 300);
-};
-
-function fadeOut() {
-    $('h1').fadeOut(fadeTime);
-    $('h2').fadeOut(fadeTime);
-};
-
-function fadeOutAlbum() {
-    $('#bg').fadeOut(fadeTime);
-};
-
-function fadeInAlbum() {
-    $('#bg').fadeIn(fadeTime);
-};
-
-function hideHeader() {
-    $("header").fadeOut(fadeTime/2);
-    $("footer").fadeOut(fadeTime/2);
-    
-}
-function showHeader() {
-    $("header").fadeIn(fadeTime/2);
-    $("footer").fadeIn(fadeTime/2);
-}
-// Ctrl+s function for re-signin
-function doc_keyUp(e) {
-    if (e.ctrlKey && e.keyCode == 83) {
-        window.location.replace('http://localhost:8080/sign-in');
-    }
-};
-document.addEventListener('keyup', doc_keyUp, false);
-
-
-$(document).on('input', '#myRange', function() {
-    volume_control($(this).val())
-});
