+++ main.js
@@ -30,7 +30,7 @@ function createWindow () {
     backgroundColor: '#000000',
     webPreferences: {
         nodeIntegration: true,
-        devTools: false
+        devTools: true,
     },
     frame: false
   });
